DRY up the two login controllers with a trait
